signed/unsigned agreement for image size in png.c

+++ protocol/fractal/utils/png.c
@@ -83,19 +83,19 @@ int bmp_to_png(unsigned char* bmp, unsigned int size, AVPacket* pkt) {
         LOG_ERROR("BMP header BM failed");
         return -1;
     }
-    unsigned pixeloffset = *((int*)(&bmp[10]));
+    int pixeloffset = *((int*)(&bmp[10]));
     w = *((int*)(&bmp[18]));
     h = *((int*)(&bmp[22]));
     if (bmp[28] != 24 && bmp[28] != 32) {
         LOG_ERROR("BMP is not 3 or 4 channel");
         return -1;
     }
-    unsigned numChannels = bmp[28] / 8;
+    int numChannels = bmp[28] / 8;
 
     // BMP pixel arrays are always multiples of 4. Images with widths that are not multiples
     //  of 4 are always padded at the end of each row. scanlineBytes is the padded width of each
     //  BMP row in the original image.
-    unsigned scanlineBytes = w * numChannels;
+    int scanlineBytes = w * numChannels;
     if (scanlineBytes % 4 != 0) scanlineBytes = (scanlineBytes / 4) * 4 + 4;
 
     int data_size = *((int*)(&bmp[34]));
@@ -105,7 +105,7 @@ int bmp_to_png(unsigned char* bmp, unsigned int size, AVPacket* pkt) {
 
     // memcpy will face UB problems below if the calculated size does not match the actual
     // BMP byte array size
-    if (size < data_size + pixeloffset) {
+    if (size < (unsigned)data_size + pixeloffset) {
         LOG_WARNING("Actual size does not match given data_size and pixeloffset (%d < %d + %d)",
                     size, data_size, pixeloffset);
         return -1;
